feat: Renamed the Onboarding screen.
