+++ Hoop_App/components/OnboardingScreen.tsx
